+++ LogicBuilder.EntityFrameworkCore.PostgreSql.Tests/Data/Configurations/ITableConfiguration.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace LogicBuilder.EntityFrameworkCore.PostgreSql.Tests.Data.Configurations
+{
+    interface ITableConfiguration
+    {
+        void Configure(ModelBuilder modelBuilder);
+    }
+}