+++ LogicBuilder.EntityFrameworkCore.PostgreSql.Tests/Data/Department.cs
@@ -1,0 +1,36 @@
+﻿using System;
+using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations;
+using System.ComponentModel.DataAnnotations.Schema;
+
+namespace LogicBuilder.EntityFrameworkCore.PostgreSql.Tests.Data
+{
+    [Table("Department")]
+    public class Department : BaseDataClass
+    {
+        [Key]
+        public int DepartmentID { get; set; }
+
+        [StringLength(50, MinimumLength = 3)]
+        public string Name { get; set; } = "";
+
+        [DataType(DataType.Currency)]
+        [Column(TypeName = "money")]
+        public decimal Budget { get; set; }
+
+        [DataType(DataType.Date)]
+        [DisplayFormat(DataFormatString = "{0:yyyy-MM-dd}", ApplyFormatInEditMode = true)]
+        [Display(Name = "Start Date")]
+        public DateTime StartDate { get; set; }
+
+        public int? InstructorID { get; set; }
+
+        [ConcurrencyCheck]
+        [DatabaseGenerated(DatabaseGeneratedOption.Computed)]
+        public uint Xmin { get; set; }
+
+        [ForeignKey("InstructorID")]
+        public Instructor? Administrator { get; set; }
+        public ICollection<Course>? Courses { get; set; }
+    }
+}