+++ LogicBuilder.EntityFrameworkCore.PostgreSql.Tests/Data/Stores/ISchoolStore.cs
@@ -1,0 +1,8 @@
+﻿using LogicBuilder.EntityFrameworkCore.Crud.DataStores;
+
+namespace LogicBuilder.EntityFrameworkCore.PostgreSql.Tests.Data.Stores
+{
+    public interface ISchoolStore : IStore
+    {
+    }
+}