+++ LogicBuilder.EntityFrameworkCore.PostgreSql.Tests/Models/Grade.cs
@@ -1,0 +1,11 @@
+﻿namespace LogicBuilder.EntityFrameworkCore.PostgreSql.Tests.Models
+{
+    public enum Grade
+    {
+		A,
+		B,
+		C,
+		D,
+		F
+    }
+}