+++ LogicBuilder.EntityFrameworkCore.PostgreSql.Tests/Models/Repositories/IDataClassesRepository.cs
@@ -1,0 +1,8 @@
+﻿using LogicBuilder.EntityFrameworkCore.Repositories;
+
+namespace LogicBuilder.EntityFrameworkCore.PostgreSql.Tests.Models.Repositories
+{
+    public interface IDataClassesRepository : IContextRepository
+    {
+    }
+}