+++ LogicBuilder.EntityFrameworkCore/Constants.cs
@@ -1,4 +1,5 @@
 ﻿[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("LogicBuilder.EntityFrameworkCore.SqlServer.IntegrationTests")]
+[assembly: System.Runtime.CompilerServices.InternalsVisibleTo("LogicBuilder.EntityFrameworkCore.PostgreSql.Tests")]
 namespace LogicBuilder.EntityFrameworkCore
 {
     public struct ExpressionOperators